+++ gomoku_rust_lib/src/lib.rs
@@ -8,7 +8,6 @@ use std::time::Instant;
 mod check;
 use check::checking_move;
 
-
 mod negamax;
 mod state;
 use state::apply_state_move;
@@ -24,10 +23,15 @@ static ALPHABET: [char; 26] = [
 ];
 
 #[pyfunction]
-fn ai_move(board: Vec<Vec<i8>>, player: i8, x: isize, y: isize) -> PyResult<((isize, isize), i32)> {
+fn ai_move(
+    board: Vec<Vec<i8>>,
+    player: i8,
+    x: isize,
+    y: isize,
+) -> PyResult<((isize, isize), isize)> {
     let mut mutboard: Vec<Vec<i8>> = board;
     let mut state: state::State = state::create_new_state(&mut mutboard, player, (x, y));
-	let start = Instant::now();
+    let start = Instant::now();
     let value = negamax::negamax(&mut state, 2, -1000, 1000, player);
     let ai_move = negamax::return_move(&mut state, value);
     let end = Instant::now();
@@ -49,8 +53,7 @@ fn place_stone(board: Vec<Vec<i8>>, player: i8, x: isize, y: isize) -> PyResult<
 
     let mut mutboard: Vec<Vec<i8>> = board;
     let mut state: state::State = state::create_new_state(&mut mutboard, player, (x, y));
-    let board_check: HashMap<String, i8> =
-        checking_move(&state);
+    let board_check: HashMap<String, i8> = checking_move(&state);
     if board_check["is_wrong_move"] == 0 {
         apply_state_move(&mut state, board_check["stone_captured"]);
         dict.set_item("board", &state.board)?;
